feat:
- change to new folder structure
- introduce interface to handler
- move Run func to server package
